Rename validator.py into validators.py
